requirements.txt
streamlit
groq
reportlab
PyMuPDF
python-docx
langdetect
spacy
